requirements.txt
streamlit==1.18.1
pandas==1.5.3
matplotlib==3.6.3
nbformat==5.7.0
nbconvert==7.2.9
Pillow==9.4.0
streamlit-option-menu==0.3.2
snowflake-connector-python==3.0.3